project/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.54
python-dotenv>=1.0.1
requests>=2.31.0
openai>=1.45.0
finvizfinance>=1.3.0
